add identity service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -45,13 +46,13 @@ namespace AuctionService.Controllers
             if(auction == null) return NotFound();
             return _mapper.Map<AuctionDto>(auction);
         }
-    
+
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            //TODO: add current user as seller
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name; ;
             _context.Auctions.Add(auction);
 
             //publish the created auction
@@ -64,13 +65,15 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
             var auction = await _context.Auctions.Include(x=>x.Item)
                 .FirstOrDefaultAsync(x=>x.Id == id);
             if(auction == null) return NotFound();
-            //TODO: check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -86,12 +89,14 @@ namespace AuctionService.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
-            //TODO: check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             _context.Auctions.Remove(auction);
 
             //publish the deleted auction
